pyproject.toml
[tool.poetry]
name = "chatgpt-line-bot"
version = "0.1.0"
description = ""
authors = ["kamibayashi <kamibayashi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
openai = "^0.27.2"
line-bot-sdk = "^2.4.2"
aiolinebot = "^0.4.1"
langchain = "^0.0.118"
transformers = "^4.27.2"
python-json-logger = "^2.0.7"
sentry-sdk = "^1.17.0"
mysql-connector-python = "^8.0.32"
sqlalchemy = "1.4.47"
python-ulid = "^1.1.0"
ipython = "^8.11.0"
alembic = "^1.10.2"
poethepoet = "^0.19.0"
isort = "^5.12.0"
pre-commit = "^3.2.0"
pytest = "^7.2.2"
toml = "^0.10.2"

[tool.poe.tasks]
test = { "cmd" = "pytest", help = "pytest" }
makemigrations = { "shell" = "alembic revision --autogenerate", help = "make db migration files." }
migrate = { "shell" = "alembic upgrade heads", help = "migrate db." }
pre-commit = { "cmd" = "pre-commit", help = "pre-commit" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
